refactor: BlogForm으로 전해주던 props 통합
